Update type hints for future optional dependencies

+++ ConservedWaterSearch/water_clustering.py
@@ -1,15 +1,17 @@
 from __future__ import annotations
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+    from matplotlib.figure import Figure
+    from nglview import NGLWidget
+    from io import TextIOWrapper
+    from hdbscan import HDBSCAN
 
 import os
 import warnings
-from io import TextIOWrapper
 
-import hdbscan
-import matplotlib.pyplot as plt
 import numpy as np
-from matplotlib.axes import Axes
-from matplotlib.figure import Figure
-from nglview import NGLWidget
 from sklearn.cluster import OPTICS, cluster_optics_xi
 
 from ConservedWaterSearch.hydrogen_orientation import (
@@ -290,7 +292,8 @@ class WaterClustering:
                 np.savetxt("water_coords_restart1.dat", np.c_[Odata, H1, H2])
             else:
                 np.savetxt("water_coords_restart1.dat", np.c_[Odata])
-            os.rename("water_coords_restart1.dat", "water_coords_restart.dat")
+            if os.path.isfile("water_coords_restart1.dat"):
+                os.rename("water_coords_restart1.dat", "water_coords_restart.dat")
         return Odata, H1, H2
 
     def _add_water_solutions(
@@ -412,16 +415,19 @@ class WaterClustering:
             # loop over minsamps- from N(snapshots) to 0.75*N(snapshots)
             for i in minsamps:
                 if clustering_algorithm == "OPTICS":
-                    clust: OPTICS | hdbscan.HDBSCAN = OPTICS(min_samples=int(i), n_jobs=self.njobs)  # type: ignore
+                    clust: OPTICS | HDBSCAN = OPTICS(min_samples=int(i), n_jobs=self.njobs)  # type: ignore
                     clust.fit(Odata)
                 # loop over xi
                 for j in lxis:
                     # recalculate reachability - OPTICS reachability has to be recaculated when changing minsamp
                     if clustering_algorithm == "HDBSCAN":
+                        try:
+                            import hdbscan
+                        except:
+                            raise Exception("install hdbscan")
+
                         clust = hdbscan.HDBSCAN(
-                            min_cluster_size=int(
-                                self.nsnaps * self.numbpct_oxygen
-                            ),
+                            min_cluster_size=int(self.nsnaps * self.numbpct_oxygen),
                             min_samples=int(i),
                             max_cluster_size=int(
                                 self.nsnaps * (2 - self.numbpct_oxygen)
@@ -464,6 +470,11 @@ class WaterClustering:
                         whichH=whichH,
                     )
                     if self.debugO == 1:
+                        try:
+                            import matplotlib.pyplot as plt
+                        except:
+                            raise Exception("install matplotlib")
+
                         plt.close(ff)
                     if len(waters) > 0:
                         found = True
@@ -476,14 +487,21 @@ class WaterClustering:
                                 fname="Clustering_results_temp1.dat",
                                 typefname="Type_Clustering_results_temp1.dat",
                             )
-                            os.rename(
-                                "Clustering_results_temp1.dat",
-                                "Clustering_results_temp.dat",
-                            )
-                            os.rename(
-                                "Type_Clustering_results_temp1.dat",
-                                "Type_Clustering_results_temp.dat",
-                            )
+                            if os.path.isfile(
+                                "Clustering_results_temp1.dat"
+                            ) and os.path.isfile("Type_Clustering_results_temp1.dat"):
+                                os.rename(
+                                    "Clustering_results_temp1.dat",
+                                    "Clustering_results_temp.dat",
+                                )
+                                os.rename(
+                                    "Type_Clustering_results_temp1.dat",
+                                    "Type_Clustering_results_temp.dat",
+                                )
+                            else:
+                                raise Warning(
+                                    "unable to overwrite temp save files. Restarting might not work properly"
+                                )
                         break
                 if found:
                     break
@@ -498,6 +516,11 @@ class WaterClustering:
             if len(Odata) < self.nsnaps:
                 found = False
         if (self.debugH == 1 or self.debugO == 1) and self.plotend:
+            try:
+                import matplotlib.pyplot as plt
+            except:
+                raise Exception("install matplotlib")
+
             plt.show()
         if self.save_results_after_done:
             self.save_results()
@@ -537,10 +560,7 @@ class WaterClustering:
                 allowed, or "onlyO" for oxygen clustering only.
                 Defaults to ["FCW", "HCW", "WCW"].
         """
-        if (
-            clustering_algorithm != "OPTICS"
-            and clustering_algorithm != "HDBSCAN"
-        ):
+        if clustering_algorithm != "OPTICS" and clustering_algorithm != "HDBSCAN":
             raise Exception("clustering algorithm must be OPTICS or HDBSCAN")
         for i in whichH:
             if not (i in ["FCW", "HCW", "WCW", "onlyO"]):
@@ -567,7 +587,7 @@ class WaterClustering:
                 whichH,
             )
         if clustering_algorithm == "OPTICS":
-            clust: OPTICS | hdbscan.HDBSCAN = OPTICS(
+            clust: OPTICS | HDBSCAN = OPTICS(
                 min_samples=minsamp,
                 min_cluster_size=int(self.numbpct_oxygen * self.nsnaps),
                 xi=xi,
@@ -576,7 +596,12 @@ class WaterClustering:
             clust.fit(Odata)
             clusters = clust.labels_
         if clustering_algorithm == "HDBSCAN":
-            clust: OPTICS | hdbscan.HDBSCAN = hdbscan.HDBSCAN(
+            try:
+                import hdbscan
+            except:
+                raise Exception("install hdbscan")
+
+            clust: OPTICS | HDBSCAN = hdbscan.HDBSCAN(
                 min_cluster_size=int(self.nsnaps * self.numbpct_oxygen),
                 min_samples=minsamp,
                 max_cluster_size=int(self.nsnaps * (2 - self.numbpct_oxygen)),
@@ -609,6 +634,11 @@ class WaterClustering:
             whichH=whichH,
         )
         if self.debugO == 1:
+            try:
+                import matplotlib.pyplot as plt
+            except:
+                raise Exception("install matplotlib")
+
             plt.close(ff)
         self._add_water_solutions(waters)
         if self.save_intermediate_results:
@@ -616,14 +646,20 @@ class WaterClustering:
                 fname="Clustering_results_temp1.dat",
                 typefname="Type_Clustering_results_temp1.dat",
             )
-            os.rename(
-                "Clustering_results_temp1.dat", "Clustering_results_temp.dat"
-            )
-            os.rename(
-                "Type_Clustering_results_temp1.dat",
-                "Type_Clustering_results_temp.dat",
-            )
+            if os.path.isfile("Clustering_results_temp1.dat") and os.path.isfile(
+                "Type_Clustering_results_temp1.dat"
+            ):
+                os.rename("Clustering_results_temp1.dat", "Clustering_results_temp.dat")
+                os.rename(
+                    "Type_Clustering_results_temp1.dat",
+                    "Type_Clustering_results_temp.dat",
+                )
         if (self.debugH == 1 or self.debugO == 1) and self.plotend:
+            try:
+                import matplotlib.pyplot as plt
+            except:
+                raise Exception("install matplotlib")
+
             plt.show()
         if self.save_results_after_done:
             self.save_results()
@@ -713,6 +749,11 @@ class WaterClustering:
                         self.normalize_orientations,
                     )
                     if self.plotreach and self.debugH > 0:
+                        try:
+                            import matplotlib.pyplot as plt
+                        except:
+                            raise Exception("install matplotlib")
+
                         plt.show()
                     if len(hyd) > 0:
                         # add water atoms for pymol visualisation
@@ -727,6 +768,11 @@ class WaterClustering:
                             and self.plotreach == 0
                             and self.debugH == 0
                         ):
+                            try:
+                                import matplotlib.pyplot as plt
+                            except:
+                                raise Exception("install matplotlib")
+
                             plt.show()
                         if stop_after_frist_water_found:
                             return waters, np.argwhere(clusters == k)
@@ -792,9 +838,7 @@ class WaterClustering:
         if len(self._waterH1) == 0 and len(self._waterH2) == 0:
             np.savetxt(fname, np.c_[self._waterO])
         else:
-            np.savetxt(
-                fname, np.c_[self._waterO, self._waterH1, self._waterH2]
-            )
+            np.savetxt(fname, np.c_[self._waterO, self._waterH1, self._waterH2])
         np.savetxt(typefname, np.c_[self._water_type], fmt="%s")
 
     def restart_cluster(
@@ -1146,6 +1190,11 @@ def __oxygen_clustering_plot(
     if type(cc) != OPTICS:
         plotreach = False
     if debugO > 0:
+        try:
+            import matplotlib.pyplot as plt
+        except:
+            raise Exception("install matplotlib")
+
         fig: Figure = plt.figure()
         if plotreach:
             ax: Axes = fig.add_subplot(1, 2, 1, projection="3d")
@@ -1169,6 +1218,11 @@ def __oxygen_clustering_plot(
         ax.set_title(title)
         ax.legend()
         if plotreach:
+            try:
+                import matplotlib.pyplot as plt
+            except:
+                raise Exception("install matplotlib")
+
             lblls = cc.labels_[cc.ordering_]
             ax = fig.add_subplot(1, 2, 2)
             plt.gca().set_prop_cycle(None)
@@ -1190,5 +1244,10 @@ def __oxygen_clustering_plot(
                     )
             ax.legend()
     if debugO == 2:
+        try:
+            import matplotlib.pyplot as plt
+        except:
+            raise Exception("install matplotlib")
+
         plt.show()
     return fig
